# swagger.json
{
  "swagger": "2.0",
  "info": {},
  "paths": {
    "/app/brand/basic": {
      "post": {
        "description": "This endpoint will return brand basic information.\n\nEndpoint: /app/brand/basic\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Brand"
        ],
        "summary": "getBrandsBasic",
        "operationId": "getBrandsBasic",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/GetBrandsByIDBasicOpts"
            }
          },
          {
            "description": "Customer login required for successful response.",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "description": "Token required for successful response.",
            "name": "auth token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " OK",
            "schema": {
              "$ref": "#/definitions/GetBrandBasicESEesp"
            }
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
    "/app/brand/username/basic": {
      "post": {
        "description": "This endpoint will return brand basic detail by username.\n\nEndpoint: /app/brand/username/basic\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Brand"
        ],
        "summary": "getBrandsBasicByUsername",
        "operationId": "getBrandsBasicByUsername",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/GetBrandsByUsernameBasicOpts"
            }
          },
          {
            "description": "Customer login required for successful response.",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "description": "Token required for successful response.",
            "name": "auth token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " OK",
            "schema": {
              "$ref": "#/definitions/GetBrandBasicESEesp"
            }
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
    "/app/brand/username/{username}": {
      "get": {
        "description": "This endpoint will return brand information by the username of the brand.\n\nEndpoint: /app/brand/username/{username}\n\nMethod: GET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Brand"
        ],
        "summary": "getBrandInfoByUsername",
        "operationId": "getBrandInfoByUsername",
        "parameters": [
          {
            "enum": [
              "vasu_pal_1"
            ],
            "type": "string",
            "name": "username",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/GetBrandsInfoByUsernameOpts"
            }
          },
          {
            "description": "Customer login required for successful response.",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "description": "Token required for successful response.",
            "name": "auth token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " OK",
            "schema": {
              "$ref": "#/definitions/GetBrandInfoEsResp"
            }
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
    "/app/brand/{brandID}": {
      "get": {
        "description": "This endpoint will return information of one brand.\n\nEndpoint: /app/brand/{brandID}\n\nMethod: GET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Brand"
        ],
        "summary": "getBrandInfo",
        "operationId": "getBrandInfo",
        "parameters": [
          {
            "enum": [
              "6065d4503824bf77961c21ae"
            ],
            "type": "object",
            "name": "brandID",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/GetBrandsInfoByIDOpts"
            }
          },
          {
            "description": "Customer login required for successful response.",
            "name": "cookie",
            "in": "header",
            "required": true
          },
          {
            "description": "Token required for successful response.",
            "name": "auth token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " OK",
            "schema": {
              "$ref": "#/definitions/GetBrandInfoEsResp"
            }
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
    "/app/brands": {
      "get": {
        "description": "This endpoint will return active brand list.\n\nEndpoint: /app/brands\n\nMethod: GET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Brand"
        ],
        "summary": "getActiveBrandsList",
        "operationId": "getActiveBrandsList",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/GetBrandsListOpts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": " OK",
            "schema": {
              "$ref": "#/definitions/GetActiveBrandsListESEesp"
            }
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
...
    "/brand/check/username": {
      "get": {
        "description": "This endpoint will check the brand username exists or not.\n\nEndpoint: /brand/check/username\n\nMethod: GET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "CheckUsername"
        ],
        "summary": "checkBrandUsernameExists",
        "operationId": "checkBrandUsernameExists",
        "parameters": [
          {
            "enum": [
              "falthead"
            ],
            "type": "string",
            "name": "username",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " true"
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
    "/brand/user/forgot-password": {
      "post": {
        "description": "This endpoint will help the brand to recover forget password.\n\nEndpoint: /brand/user/forgot-password\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "BrandDashAPIs"
        ],
        "summary": "brandUserForgotPassword",
        "operationId": "brandUserForgotPassword",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ForgotPasswordOpts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": " true"
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
    "/brand/user/login": {
      "post": {
        "description": "This endpoint will login the brand user.\n\nEndpoint: /brand/user/login\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "BrandDashAPIs"
        ],
        "summary": "brandUserLogin",
        "operationId": "brandUserLogin",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/BrandUserLoginOpts"
            }
          },
          {
            "type": "string",
            "name": "returnToken",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " token"
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
    "/brand/user/reset-password": {
      "post": {
        "description": "This endpoint will reset the brand password.\n\nEndpoint: /brand/user/reset-password\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "BrandDashAPIs"
        ],
        "summary": "brandUserResetPassword",
        "operationId": "brandUserResetPassword",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ResetPasswordOpts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": " true"
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
    "/brand/{brandID}": {
      "get": {
        "description": "This endpoint will return brand details.\n\nEndpoint: /brand/{brandID}\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Brand"
        ],
        "summary": "getBrandByID",
        "operationId": "getBrandByID",
        "parameters": [
          {
            "enum": [
              "60b50277a97a2d73b211aec7"
            ],
            "type": "object",
            "name": "brandID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": " Success",
            "schema": {
              "$ref": "#/definitions/getBrandByID"
            }
          },
          "400": {
            "description": " Error",
            "schema": {
              "$ref": "#/definitions/CommonError"
            }
          }
        }
      }
    },
...
    "/user/forgot-password": {
      "post": {
        "description": "This endpoint will help the user to recover the password.\n\nEndpoint: /user/forgot-password\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Password"
        ],
        "summary": "forgotPassword",
        "operationId": "forgotPassword",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ForgotPasswordOpts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": " true"
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
...
    "/user/reset-password": {
      "post": {
        "description": "This endpoint will help the user reset the password.\n\nEndpoint: /user/reset-password\n\nMethod: POST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Password"
        ],
        "summary": "resetPassword",
        "operationId": "resetPassword",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/ResetPasswordOpts"
            }
          }
        ],
        "responses": {
          "200": {
            "description": " true"
          },
          "400": {
            "description": " BadRequest",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          },
          "403": {
            "description": "Invalid User",
            "schema": {
              "$ref": "#/definitions/AppErr"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AppErr": {
      "type": "object",
      "properties": {
        "Error": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "RequestID": {
          "type": "string"
        }
      },
      "x-go-package": "go-app/server/handler"
    },
...
    "GetActiveBrandsListESEesp": {
      "type": "object",
      "properties": {
        "id": {
          "$ref": "#/definitions/ObjectID"
        },
        "logo": {
          "$ref": "#/definitions/IMG"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "go-app/schema"
    },
    "GetBrandBasicESEesp": {
      "type": "object",
      "properties": {
        "id": {
          "$ref": "#/definitions/ObjectID"
        },
        "is_followed_by_user": {
          "type": "boolean",
          "x-go-name": "IsFollowedByUser"
        },
        "logo": {
          "$ref": "#/definitions/IMG"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "go-app/schema"
    },
...
    "GetBrandsByIDBasicOpts": {
      "type": "object",
      "properties": {
        "CustomerID": {
          "$ref": "#/definitions/ObjectID"
        },
        "ids": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "IDs"
        }
      },
      "x-go-package": "go-app/schema"
    },
    "GetBrandsByUsernameBasicOpts": {
      "type": "object",
      "properties": {
        "CustomerID": {
          "$ref": "#/definitions/ObjectID"
        },
        "usernames": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Usernames"
        }
      },
      "x-go-package": "go-app/schema"
    },
    "GetBrandsInfoByIDOpts": {
      "type": "object",
      "properties": {
        "id": {
          "$ref": "#/definitions/ObjectID"
        },
        "user_id": {
          "$ref": "#/definitions/ObjectID"
        }
      },
      "x-go-package": "go-app/schema"
    },
    "GetBrandsInfoByUsernameOpts": {
      "type": "object",
      "properties": {
        "user_id": {
          "$ref": "#/definitions/ObjectID"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "go-app/schema"
    },
    "GetBrandsListOpts": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Page"
        },
        "size": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Size"
        }
      },
      "x-go-package": "go-app/schema"
    },
...
  }
}